refactor: input 검증 사항 책임을 view에게 전임
