+++ src/main/java/lotto/LottoGame.java
@@ -7,7 +7,6 @@ import lotto.dto.WinningResults;
 import lotto.generator.RandomLottoNumberGenerator;
 import lotto.generator.WinningResultMessageGenerator;
 import lotto.resolver.ExceptionResolver;
-import lotto.validator.input.InputCommonValidator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -30,45 +29,21 @@ public class LottoGame {
         printWinningResult();
     }
 
-    private void buyLottos() {
-        int price = inputBuyingPrice();
+    public void buyLottos() {
+        int price = ExceptionResolver.resolveInput(inputView::inputBuyingPrice);
         lottoMachine.buyLottos(RandomLottoNumberGenerator.getSupplier(), price);
     }
 
-    private int inputBuyingPrice() {
-        String inputPrice = inputView.inputBuyingPrice();
-        InputCommonValidator.validateSingle(inputPrice);
-        return Integer.parseInt(inputPrice);
-    }
-
     private void inputWinningLotto() {
-        List<Integer> winningNumbers = ExceptionResolver.resolveInput(this::inputLottoNumbers);
-        int bonusNumber = ExceptionResolver.resolveInput(this::inputBonusNumber);
+        List<Integer> winningNumbers = ExceptionResolver.resolveInput(inputView::inputLottoNumbers);
+        int bonusNumber = ExceptionResolver.resolveInput(inputView::inputBonusNumber);
         lottoMachine.addWinningLotto(winningNumbers, bonusNumber);
-    }
-
-    private List<Integer> inputLottoNumbers() {
-        List<String> inputNumbers = inputView.inputLottoNumbers();
-        InputCommonValidator.validateMultiple(inputNumbers);
-        return convertToNumbers(inputNumbers);
-    }
-
-    private int inputBonusNumber() {
-        String inputNumber = inputView.inputBonusNumber();
-        InputCommonValidator.validateSingle(inputNumber);
-        return Integer.parseInt(inputNumber);
     }
 
     private void printWinningResult() {
         WinningResults winningResults = lottoMachine.createWinningResults();
         String message = WinningResultMessageGenerator.generate(winningResults);
         outputView.printWinningResult(message);
-    }
-
-    private List<Integer> convertToNumbers(final List<String> inputNumbers) {
-        return inputNumbers.stream()
-                .map(Integer::parseInt)
-                .toList();
     }
 
     private void printBuyingResults() {
