+++ src/main/java/lotto/view/OutputView.java
@@ -2,8 +2,8 @@ package lotto.view;
 
 public class OutputView {
 
-    public static final String BUYING_COUNT_FORMAT = "%d개를 구매했습니다.";
-    public static final String RESULT_STATISTICS = """
+    private static final String BUYING_COUNT_FORMAT = "%d개를 구매했습니다.";
+    private static final String RESULT_STATISTICS = """
             당첨 통계
             ---""";
